Imported Upstream version 2.1.2+dfsg

+++ src/modules/grid/grid_gridding/Interpolation_NaturalNeighbour.cpp
@@ -1,5 +1,5 @@
 /**********************************************************
- * Version $Id: Interpolation_NaturalNeighbour.cpp 1921 2014-01-09 10:24:11Z oconrad $
+ * Version $Id: Interpolation_NaturalNeighbour.cpp 2064 2014-03-21 13:20:57Z oconrad $
  *********************************************************/
 
 ///////////////////////////////////////////////////////////
@@ -135,9 +135,9 @@ bool CInterpolation_NaturalNeighbour::Interpolate(void)
 	delaunay	*pTIN	= delaunay_build(n, pSrc, 0, NULL, 0, NULL);
 
 	//-----------------------------------------------------
-	double	*xDst	= (double *)SG_Malloc((long) m_pGrid->Get_NCells() * sizeof(double));
-	double	*yDst	= (double *)SG_Malloc((long) m_pGrid->Get_NCells() * sizeof(double));
-	double	*zDst	= (double *)SG_Malloc((long) m_pGrid->Get_NCells() * sizeof(double));
+	double	*xDst	= (double *)SG_Malloc(m_pGrid->Get_NCells() * sizeof(double));
+	double	*yDst	= (double *)SG_Malloc(m_pGrid->Get_NCells() * sizeof(double));
+	double	*zDst	= (double *)SG_Malloc(m_pGrid->Get_NCells() * sizeof(double));
 
 	for(y=0, i=0, p.y=m_pGrid->Get_YMin(); y<m_pGrid->Get_NY() && Set_Progress(y, m_pGrid->Get_NY()); y++, p.y+=m_pGrid->Get_Cellsize())
 	{
